fixed to get the /api/v1/submit endpoint working

+++ crates/db/src/models/submissions.rs
@@ -20,7 +20,7 @@ pub struct Submissions {
     is_banned: bool,
     deleted_utc: i64,
     distinguish_level: i16,
-    gm_distinguish: i16,
+    gm_distinguish: i32,
     stickied: bool,
     is_approved: i32,
     approved_utc: i64,
